+++ files/gitlab-cookbooks/gitlab/libraries/helpers/base_pg_helper.rb
@@ -13,6 +13,10 @@ class BasePgHelper
     OmnibusHelper.new(node).service_up?(service_name)
   end
 
+  def is_managed_and_offline?
+    OmnibusHelper.new(node).is_managed_and_offline?(service_name)
+  end
+
   def database_exists?(db_name)
     psql_cmd(["-d 'template1'",
               "-c 'select datname from pg_database' -A",